Place fullscreen and volume controls beside play/pause
