+++ src/components/VideoPlayer.css
@@ -25,44 +25,3 @@
   left: 0;
   object-fit: cover;
 }
-
-.video-hover-controls {
-  position: absolute;
-  bottom: 8px;
-  right: 8px;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  opacity: 0;
-  transition: opacity 0.3s ease;
-  pointer-events: none;
-}
-
-.video-hover-controls.visible {
-  opacity: 1;
-  pointer-events: auto;
-}
-
-.volume-control {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  background: rgba(0, 0, 0, 0.5);
-  padding: 4px;
-  border-radius: 4px;
-}
-
-.volume-control input[type='range'] {
-  width: 80px;
-}
-
-.fullscreen-btn {
-  background: rgba(0, 0, 0, 0.5);
-  border: none;
-  padding: 4px;
-  border-radius: 4px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
